Made navbar reuseable and global
