+++ app/category/[category]/page.tsx
@@ -56,7 +56,7 @@
       favorites.some((f) => f.idMeal === idMeal);
 
     return (
-      <div className="flex flex-col items-center justify-center p-5 md:p-10 w-full min-h-screen bg-base-100">
+      <div className="flex flex-col items-center justify-center p-5 md:p-10 w-full bg-base-100 min-h-screen py-10 px-4 relative">
         <BackButton />
         <h1 className="text-4xl md:text-6xl text-secondary mb-10 capitalize">
           {params.category} 🍽
